Refactor TransportAlterRoleAction
similar to https://github.com/crate/crate/pull/15618
